fixed invisibility
